feat: plugin-based judging (#2)
* feat(server): plugin-based judging with contest type, evaluator, and checker format handlers

* fix: revert accidental regressions from plugin-based judging merge

* feat(sdk): server sdk

* fix(test): integration test has empty registries

* feat(sdk): modularize checker formats

* feat(sdk): trait-based plugin host to allow mocking

* feat(cli): scaffold and build plugin

* fix(worker): worker performance + isolate abstraction bugs

* feat(plugin): add resource-level plugin config support

* feat(plugin): standard checkers plugin

* build(server-sdk): add metadata for crates.io publishing

* chore(plugin): use crates.io server sdk instead of local workspace one

* feat(plugin): add config schema declarations and admin config ui

* feat(plugin): add batch evaluator plugin and beautify plugin config ui

* fix(plugin): make batch evaluator resource limits configurable

* feat(web): ui for resource-level plugin config

* feat(web): shared-deps vite plugin + dev-mode plugin loading + config ui fixes

* feat(web): expose contest type, problem type, and checker format to the ui

* feat(server,web): support standalone problem submissions with default contest type setting

* refactor(server): split handlePluginRequest into multiple public apis

* feat(plugin,cli): hot reload plugins with watch mode

* feat: change score fields from i32 to f64 across entities and judge pipeline

* feat(hooks): overhaul hook system with scoped/blocking/notify modes

* feat(server-sdk): incremental submission updates, hook events, stdout/stderr capture

* feat(server): hook dispatch, plugin config namespacing, and entity updates

* feat(web): plugin rejection UI, running status, shared-deps expansion

* feat(plugins): add cooldown, submission-limit, and IOI plugins

* fix(plugin): enhance hook system

* fix(ioi-plugin): enhance ioi plugin

* fix(common,server,worker): make clippy happy

* fix(server): add label to tests

* fix(plugin): bugs

* fix(web0: submissionresult taking over plugin injected hook when running submission

* refactor(ioi-plugin): simplify scoring config logic and show static layout in running-state renderer instead of incrementally changing it

* fix(cli): only upload on plugin package fingerprint change

* fix(ioi-plugin): properly align ioi contestinfo and submissionresult

* chore(cli): prepare cli for publishing

* fix(server): fix authz issues

* feat(web): beautify device auth page

* refactor(server-sdk): move host fn types to sdk

* fix(cli): break the watch infinite loop

* refactor(server-sdk): extract common usecases to sdk

* refactor(plugin): standardize plugin project structure

* feat(web): show plugin config as sheet

* feat(plugin): serve plugin css via link tags

* feat(plugin): support tailwind in plugins

* refactor(plugin): migrate frontend components from inline styles to tailwind

* chore(server): submission query optimization

* feat(web): add submission detail view with test case results

* fix(sdk): harden sdk

* feat(server): support additional_files + config file uploads + shared blob store factory with worker

* feat(worker): configurable blob store backend + actually support fifo properly

* fix(plugin): native test support

* feat(communication-evaluator): communication evaluator plugin

* fix(web,worker): compile errors

* chore(worker): make clippy happy

* chore(build): make clippy happier

* chore(build): make clippy even happier

+++ packages/common/src/judge_job.rs
@@ -22,7 +22,7 @@ pub struct TestCaseData {
     /// Expected output for comparison
     pub expected_output: String,
     /// Maximum score for this test case
-    pub score: i32,
+    pub score: f64,
 }
 
 /// A judge job message sent to the worker queue.
@@ -36,7 +36,7 @@ pub struct JudgeJob {
     pub problem_id: i32,
     /// Source files to judge
     pub files: Vec<JudgeFile>,
-    /// Programming language (e.g., "cpp", "java", "python")
+    /// Programming language (e.g., "cpp", "java", "python3")
     pub language: String,
     /// Time limit in milliseconds
     pub time_limit: i32,
